+++ eod/elements/elements.go
@@ -3,7 +3,6 @@ package elements
 import (
 	"github.com/Nv7-Github/Nv7Haven/eod/base"
 	"github.com/Nv7-Github/Nv7Haven/eod/polls"
-	"github.com/Nv7-Github/Nv7Haven/eod/types"
 	"github.com/Nv7-Github/sevcord/v2"
 	"github.com/jmoiron/sqlx"
 )
@@ -28,14 +27,6 @@ func NewElements(s *sevcord.Sevcord, db *sqlx.DB, base *base.Base, polls *polls.
 			AutoComplete(e.Autocomplete),
 	))
 	s.RegisterSlashCommand(sevcord.NewSlashCommand(
-		"inv",
-		"View your inventory!",
-		e.Inv,
-		sevcord.NewOption("user", "The user to view the inventory of!", sevcord.OptionKindUser, false),
-		sevcord.NewOption("sort", "The sort order of the inventory!", sevcord.OptionKindString, false).
-			AddChoices(types.Sorts...),
-	))
-	s.RegisterSlashCommand(sevcord.NewSlashCommand(
 		"hint",
 		"Learn how to make an element!",
 		e.Hint,
@@ -49,6 +40,5 @@ func NewElements(s *sevcord.Sevcord, db *sqlx.DB, base *base.Base, polls *polls.
 		sevcord.NewOption("result", "The result of the combination!", sevcord.OptionKindString, true),
 		sevcord.NewOption("autocapitalize", "Whether or not to autocapitalize!", sevcord.OptionKindBool, false),
 	))
-	s.AddButtonHandler("inv", e.InvHandler)
 	return e
 }
