+++ eod/init.go
@@ -3,6 +3,7 @@ package eod
 import (
 	"github.com/Nv7-Github/Nv7Haven/eod/base"
 	"github.com/Nv7-Github/Nv7Haven/eod/elements"
+	"github.com/Nv7-Github/Nv7Haven/eod/pages"
 	"github.com/Nv7-Github/Nv7Haven/eod/polls"
 )
 
@@ -10,5 +11,6 @@ func (b *Bot) Init() {
 	b.base = base.NewBase(b.s, b.db)
 	b.polls = polls.NewPolls(b.db, b.base, b.s)
 	b.elements = elements.NewElements(b.s, b.db, b.base, b.polls)
+	b.pages = pages.NewPages(b.base, b.db, b.s)
 	b.s.SetMessageHandler(b.messageHandler)
 }
