+++ eod/types/types.go
@@ -94,21 +94,13 @@ type Poll struct {
 	Data PollData `db:"data"`
 }
 
-// Buttons
+// Discord util
 var Sorts = []sevcord.Choice{
 	sevcord.NewChoice("ID", "id"),
 	sevcord.NewChoice("Name", "name"),
 	sevcord.NewChoice("Creator", "creator"),
 	sevcord.NewChoice("Created On", "createdon"),
 	sevcord.NewChoice("Tree Size", "treesize"),
-}
-
-// Creates btns with name prevnext|<params>
-func PageSwitchBtns(handler, params string) []sevcord.Component {
-	return []sevcord.Component{
-		sevcord.NewButton("", sevcord.ButtonStylePrimary, handler, "prev|"+params).WithEmoji(sevcord.ComponentEmojiCustom("leftarrow", "861722690813165598", false)),
-		sevcord.NewButton("", sevcord.ButtonStylePrimary, handler, "next|"+params).WithEmoji(sevcord.ComponentEmojiCustom("rightarrow", "861722690926936084", false)),
-	}
 }
 
 // Consts
